crc32 should cover fields other than crc32 itself

+++ include/bitcask/CRC32.h
@@ -9,9 +9,9 @@
 #include <string>
 
 struct CRC32 {
-    static size_t getCRC32(const std::string &s);
+    static uint32_t getCRC32(const std::string &s);
 
-    static size_t getCRC32(char *buff, size_t len);
+    static uint32_t getCRC32(char *buff, size_t len);
 };
 
 #endif //SRC_BITCASK_CRC32_H
